+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/account/view/HistoryView.kt
@@ -306,7 +306,7 @@ private fun ComposableHistoryListItemsPreview() {
 const val ADDRESS_IN_TITLE_WIDTH_RATIO = 0.5f
 
 @Composable
-@Suppress("LongMethod")
+@Suppress("LongMethod", "CyclomaticComplexMethod")
 private fun HistoryItem(
     transaction: TransactionUi,
     isHideBalances: Boolean,
@@ -356,6 +356,27 @@ private fun HistoryItem(
         TransactionExtendedState.RECEIVE_FAILED -> {
             typeText = stringResource(id = R.string.account_history_item_receive_failed)
             typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_receive_icon)
+            textColor = ZcashTheme.colors.historyRedColor
+            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleFailed
+        }
+
+        TransactionExtendedState.SHIELDED -> {
+            typeText = stringResource(id = R.string.account_history_item_shielded_funds)
+            typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_shielded_funds)
+            textColor = MaterialTheme.colorScheme.onBackground
+            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleRegular
+        }
+
+        TransactionExtendedState.SHIELDING -> {
+            typeText = stringResource(id = R.string.account_history_item_shielding_funds)
+            typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_shielded_funds)
+            textColor = ZcashTheme.colors.textDescription
+            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleRunning
+        }
+
+        TransactionExtendedState.SHIELDING_FAILED -> {
+            typeText = stringResource(id = R.string.account_history_item_shielding_failed)
+            typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_shielded_funds)
             textColor = ZcashTheme.colors.historyRedColor
             textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleFailed
         }
@@ -556,7 +577,7 @@ private fun HistoryItemCollapsedAddressPart(
                 )
             }
         }
-    } else {
+    } else if (!transaction.overview.isShielding) {
         Icon(
             imageVector = ImageVector.vectorResource(R.drawable.ic_trx_shielded),
             contentDescription = stringResource(id = R.string.account_history_item_shielded)
@@ -636,7 +657,7 @@ private fun HistoryItemExpandedPart(
     modifier: Modifier = Modifier
 ) {
     Column(modifier = modifier) {
-        if (transaction.messages.containsValidMemo()) {
+        if (transaction.overview.isShielding.not() && transaction.messages.containsValidMemo()) {
             Text(
                 text =
                     pluralStringResource(
@@ -662,8 +683,8 @@ private fun HistoryItemExpandedPart(
                 )
                 Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
             }
-        } else if (transaction.recipientAddressType == null ||
-            transaction.recipientAddressType == AddressType.Shielded
+        } else if (transaction.overview.isShielding.not() &&
+            (transaction.recipientAddressType == null || transaction.recipientAddressType == AddressType.Shielded)
         ) {
             Text(
                 text = stringResource(id = R.string.account_history_item_no_message),
@@ -929,41 +950,41 @@ internal enum class TransactionExtendedState {
     SEND_FAILED,
     RECEIVED,
     RECEIVING,
-    RECEIVE_FAILED;
+    RECEIVE_FAILED,
+    SHIELDED,
+    SHIELDING,
+    SHIELDING_FAILED;
 
-    fun isFailed(): Boolean = this == SEND_FAILED || this == RECEIVE_FAILED
+    fun isShielding() = this in listOf(SHIELDED, RECEIVE_FAILED, SHIELDING_FAILED)
 
-    fun isSendType(): Boolean = this == SENDING || this == SENT || this == SEND_FAILED
+    fun isFailed(): Boolean = this in listOf(SEND_FAILED, RECEIVE_FAILED, SHIELDING_FAILED)
+
+    fun isSendType(): Boolean = this in listOf(SENDING, SENT, SEND_FAILED, SHIELDED, SHIELDING_FAILED, SHIELDING)
 }
 
 private fun TransactionOverview.getExtendedState(): TransactionExtendedState {
     return when (transactionState) {
-        TransactionState.Expired -> {
-            if (isSentTransaction) {
-                TransactionExtendedState.SEND_FAILED
-            } else {
-                TransactionExtendedState.RECEIVE_FAILED
+        TransactionState.Expired ->
+            when {
+                isShielding -> TransactionExtendedState.SHIELDING_FAILED
+                isSentTransaction -> TransactionExtendedState.SEND_FAILED
+                else -> TransactionExtendedState.RECEIVE_FAILED
             }
-        }
 
-        TransactionState.Confirmed -> {
-            if (isSentTransaction) {
-                TransactionExtendedState.SENT
-            } else {
-                TransactionExtendedState.RECEIVED
+        TransactionState.Confirmed ->
+            when {
+                isShielding -> TransactionExtendedState.SHIELDED
+                isSentTransaction -> TransactionExtendedState.SENT
+                else -> TransactionExtendedState.RECEIVED
             }
-        }
 
-        TransactionState.Pending -> {
-            if (isSentTransaction) {
-                TransactionExtendedState.SENDING
-            } else {
-                TransactionExtendedState.RECEIVING
+        TransactionState.Pending ->
+            when {
+                isShielding -> TransactionExtendedState.SHIELDING
+                isSentTransaction -> TransactionExtendedState.SENDING
+                else -> TransactionExtendedState.RECEIVING
             }
-        }
 
-        else -> {
-            error("Unexpected transaction state found while calculating its extended state.")
-        }
+        else -> error("Unexpected transaction state found while calculating its extended state.")
     }
 }