+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this application adheres to [Semantic Versioning](https://semver.org/spec/v2
 - Choose server screen has been redesigned
 - Settings and Advanced Settings screens have been redesigned
 - Android `compileSdkVersion` and `targetSdkVersion` have been updated to version 35
+- Shielded transactions are properly indicated in transaction history
 
 ### Fixed
 - The issue of printing the stacktrace of errors in dialogs has been resolved